+++ manual/src/is_greater_mint.c
@@ -1,12 +1,14 @@
 #include <mint.h>
+#include <stddef.h>
 #define max(a, b) (a)<(b)?(b):(a)
 
 int is_greater_mint_in (mint *numa, mint *numb){
-  // unsigned long sizea = size_mint(numa);
-  // unsigned long sizeb = size_mint(numb);
-  unsigned long sizea = real_size_mint(numa);
-  unsigned long sizeb = real_size_mint(numb);
-  unsigned long count = max(sizea, sizeb);
+  // unsigned long sizea = real_size_mint(numa);
+  // unsigned long sizeb = real_size_mint(numb);
+  // unsigned long count = max(sizea, sizeb);
+  size_t sizea = real_size_mint(numa);
+  size_t sizeb = real_size_mint(numb);
+  size_t count = max(sizea, sizeb);
   while (count--){
     mint_cell numca = get_mint(count, 0, numa);
     mint_cell numcb = get_mint(count, 0, numb);
