+++ manual/src/load_mint.c
@@ -1,9 +1,10 @@
 #include <mint.h>
-#include <stdio.h>
+#include <stddef.h>
 
 void load_mint_from_int (int num, mint *numout){
   char *addressnum = (char*)&num;
-  int size = sizeof(int);
+  // int size = sizeof(int);
+  size_t size = sizeof(int);
   while (size--){
     set_mint(addressnum[size], size, 0, numout);
   }
@@ -12,7 +13,8 @@ void load_mint_from_int (int num, mint *numout){
 
 void load_mint_from_long (long num, mint *numout){
   char *addressnum = (char*)&num;
-  int size = sizeof(long);
+  // int size = sizeof(long);
+  size_t size = sizeof(long);
   while (size--){
     set_mint(addressnum[size], size, 0, numout);
   }
