+++ manual/src/set_mint.c
@@ -1,10 +1,12 @@
 #include <mint.h>
-#include "size_mint.h"
 
-void set_mint (mint_cell numc, unsigned long index, unsigned int seek, mint *numin){
-  unsigned int seeka = seek;
-  unsigned int seekb = 8 - seek;
-  unsigned long size = real_size_mint(numin);
+void set_mint (mint_cell numc, size_t index, size_t seek, mint *numin){
+  // unsigned int seeka = seek;
+  // unsigned int seekb = 8 - seek;
+  // unsigned long size = real_size_mint(numin);
+  size_t seeka = seek;
+  size_t seekb = 8 - seek;
+  size_t size = real_size_mint(numin);
   if (index < size){
     numin->address_beginning[index] &= ~(0xff << seeka);
     numin->address_beginning[index] |= numc << seeka;
