+++ manual/src/shift_size_mint.c
@@ -1,13 +1,14 @@
 #include <mint.h>
+#include <stddef.h>
 
-unsigned long seek_size_mint (mint *num){
-  unsigned long numc = get_mint(0, 0, num);
+size_t seek_size_mint (mint *num){
+  size_t numc = get_mint(0, 0, num);
   return numc & 7u;
 }
 
-unsigned long offset_size_mint (mint *num){
-  unsigned long inum = 0;
-  int count = sizeof(long);
+size_t offset_size_mint (mint *num){
+  size_t inum = 0;
+  int count = sizeof(size_t);
   while (count--){
     mint_cell numc = get_mint(count, 3u, num);
     inum <<= 8;
