+++ manual/src/size_mint.c
@@ -1,9 +1,10 @@
 #include <mint.h>
+#include <stddef.h>
 
-unsigned long size_mint (mint *numin){
+size_t size_mint (mint *numin){
   return (numin->address - numin->address_beginning) / sizeof(mint_cell);
 }
 
-unsigned long real_size_mint (mint *numin){
+size_t real_size_mint (mint *numin){
   return (numin->address_end - numin->address_beginning) / sizeof(mint_cell);
 }
