+++ tmp/src/make_tmp_mint.h
@@ -1,8 +1,9 @@
 #pragma once
 #include <mint.h>
+#include <stddef.h>
 
 #define make_tmp_mint(var, size)\
-unsigned long __ ## var ## _size = size;\
+size_t __ ## var ## _size = size;\
 mint_cell __ ## var ## _data[__ ## var ## _size];\
 mint __ ## var;\
 mint *var = &__ ## var;\
