+++ tmp/src/rshift_tmp_mint.h
@@ -1,14 +1,15 @@
 #pragma once
 #include <mint.h>
+#include <stddef.h>
 
 #define rshift_tmp_mint(var, numa, numb)\
 mint *__ ## var ## _rshift_argument1 = numa;\
 mint *__ ## var ## _rshift_argument2 = numb;\
-unsigned long __ ## var ## _rshift_seek =\
+size_t __ ## var ## _rshift_seek =\
 seek_size_mint(__ ## var ## _rshift_argument2);\
-unsigned long __ ## var ## _rshift_offset =\
+size_t __ ## var ## _rshift_offset =\
 offset_size_mint(__ ## var ## _rshift_argument2);\
-unsigned long __ ## var ## _rshift_size =\
+size_t __ ## var ## _rshift_size =\
 size_mint(__ ## var ## _rshift_argument1);\
 make_tmp_mint(var,\
 __ ## var ## _rshift_size +\
